+++ app/contracts/games/update_contract.rb
@@ -5,8 +5,6 @@ module Games
     config.messages.namespace = :game
 
     params do
-      optional(:source).filled(:string)
-      optional(:external_id).filled(:string)
       required(:week_id).filled(:integer)
     end
   end