+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - table view for squad views
 - precalculating of selected by teams ratio for players
+- multiple external ids for games
 
 ## [1.1.4] - 2023-11-30
 ### Added